new story for CheckboxList
